requirements.txt
# Core web framework
fastapi
uvicorn
python-multipart

# Document parsing & OCR
pdfplumber
pytesseract
pillow

# Embeddings & ML
sentence-transformers
transformers
torch
numpy

# Vector DB / retrieval
faiss-cpu

# Optional but helpful
paddleocr
opencv-python

# Utilities & notebooks
jupyter
ipython
tqdm
